pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "flask>=3.1.2",
    "python-telegram-bot>=22.3",
]
